+++ src/pages/Main.jsx
@@ -2,11 +2,10 @@ import Messages from 'components/Messages';
 import NewMessage from 'components/NewMessage/NewMessage';
 import React from 'react';
 import styled from 'styled-components';
-import Login from 'components/Login';
+
 const Main = () => {
     return (
         <Wrapper>
-            <Login />
             <MainWrapper>
                 <Messages />
                 <NewMessage />
